requirements.txt
Django==5.0
djangorestframework==3.14.0
django-filter==22.1
